Add option to manually set maintenance mode
Use `python -m pyramid_heroku.maintenance` to manually set Heroku
maintenance mode.

Also, some cleanup and refactor.
